fix algorithm for adding sum

+++ PA1/sum_baseline.c
@@ -3,13 +3,12 @@
 
 int main(int argc, char *argv[])
 {
-  int N = atoi(argv[1]);
+  long long N = atoll(argv[1]);
   long long sum = 0;
-  for (int i = 0; i < N; i++)
+  for (long long i = 0; i < N; i++)
   {
     sum += i;
   }
   printf("Sum: %lld\n",sum);
   return 0;
 }
-
